+++ src/http/oauth2.ts
@@ -25,6 +25,7 @@ export function createOAuth2HttpClient(
   options: ParamsOauth2
 ): OAuth2HttpClient {
   let tokenData: Partial<TokenDataOauth2>;
+  let refreshingTokensPromise: Promise<TokenDataOauth2>;
 
   if ('refreshToken' in options && 'accessToken' in options) {
     tokenData = {
@@ -94,7 +95,7 @@ export function createOAuth2HttpClient(
       // Refresh 10 seconds before the token is about to expire.
       // This is to prevent being just too late with a refresh of the token due to latencies
       if (Date.now() > expireTime - 10 * 1000) {
-        await authenticate({ refreshToken: tokenData.refreshToken });
+        await refreshTokens();
       }
     }
 
@@ -123,7 +124,7 @@ export function createOAuth2HttpClient(
       !originalRequest.isRetryWithRefreshedTokens
     ) {
       originalRequest.isRetryWithRefreshedTokens = true;
-      await authenticate({ refreshToken: tokenData.refreshToken });
+      await refreshTokens();
       return httpWithAuth(originalRequest);
     }
 
@@ -178,6 +179,20 @@ export function createOAuth2HttpClient(
     );
 
     return await dateAndSetTokenData(tokenResult.data);
+  }
+
+  async function refreshTokens(): Promise<TokenDataOauth2> {
+    if (refreshingTokensPromise) {
+      return refreshingTokensPromise;
+    }
+
+    refreshingTokensPromise = authenticate({ refreshToken: tokenData.refreshToken });
+
+    try {
+      return await refreshingTokensPromise;
+    } finally {
+      refreshingTokensPromise = null;
+    }
   }
 
   async function confirmMfa({